Add files via upload
Explore Pythagoras theorem using infinitesimal random paths and extend it to non right-angled triangles and N vectors. When we travel path A and path B, the outcome is the same as travelling path C. However, it is the process of doing so that is distinct. Randomness blurs the boundaries and makes them indistinguishable. with randomess, both the process and the outcome is the same So randomly travelling C is the same thing as randomly travelling A and B. Provided they are interchangable. Interchangability provides completeness and there is no ambiguity, which allows the equation to hold.